+++ router/core/transport.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/wundergraph/graphql-go-tools/v2/pkg/engine/datasource/httpclient"
+	"go.opentelemetry.io/otel/attribute"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 
 	"github.com/wundergraph/cosmo/router/pkg/metric"
@@ -26,7 +27,6 @@ import (
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	otrace "go.opentelemetry.io/otel/trace"
 	"go.uber.org/zap"
-	"golang.org/x/sync/singleflight"
 )
 
 type TransportPreHandler func(req *http.Request, ctx RequestContext) (*http.Request, *http.Response)
@@ -39,7 +39,15 @@ type CustomTransport struct {
 	metricStore  metric.Provider
 	logger       *zap.Logger
 
-	sf *singleflight.Group
+	sf   map[uint64]*sfCacheItem
+	sfMu *sync.RWMutex
+}
+
+type sfCacheItem struct {
+	loaded   chan struct{}
+	response *http.Response
+	body     []byte
+	err      error
 }
 
 func NewCustomTransport(
@@ -59,7 +67,8 @@ func NewCustomTransport(
 		ct.roundTripper = roundTripper
 	}
 	if enableSingleFlight {
-		ct.sf = &singleflight.Group{}
+		ct.sf = make(map[uint64]*sfCacheItem)
+		ct.sfMu = &sync.RWMutex{}
 	}
 
 	return ct
@@ -68,9 +77,10 @@ func NewCustomTransport(
 func (ct *CustomTransport) measureSubgraphMetrics(req *http.Request) func(err error, resp *http.Response) {
 
 	reqContext := getRequestContext(req.Context())
-	baseFields := getAttributesFromOperationContext(reqContext.operation)
-
 	activeSubgraph := reqContext.ActiveSubgraph(req)
+
+	var baseFields []attribute.KeyValue
+
 	if activeSubgraph != nil {
 		baseFields = append(baseFields, otel.WgSubgraphName.String(activeSubgraph.Name))
 		baseFields = append(baseFields, otel.WgSubgraphID.String(activeSubgraph.Id))
@@ -79,6 +89,8 @@ func (ct *CustomTransport) measureSubgraphMetrics(req *http.Request) func(err er
 	if attributes := baseAttributesFromContext(req.Context()); attributes != nil {
 		baseFields = append(baseFields, attributes...)
 	}
+
+	baseFields = append(baseFields, reqContext.operation.Attributes()...)
 
 	inFlightDone := ct.metricStore.MeasureInFlight(req.Context(), baseFields...)
 	ct.metricStore.MeasureRequestSize(req.Context(), req.ContentLength, baseFields...)
@@ -162,11 +174,6 @@ func (ct *CustomTransport) RoundTrip(req *http.Request) (resp *http.Response, er
 	return resp, err
 }
 
-type responseWithBody struct {
-	res  *http.Response
-	body []byte
-}
-
 func (ct *CustomTransport) isUpgradeError(req *http.Request, res *http.Response) bool {
 	return req.Header.Get("Upgrade") != "" && res.StatusCode != http.StatusSwitchingProtocols
 }
@@ -214,55 +221,83 @@ func releaseBuffer(buf *bytes.Buffer) {
 
 func (ct *CustomTransport) roundTripSingleFlight(req *http.Request) (*http.Response, error) {
 
-	// We need to use the single flight group to ensure that the request is only sent once
-	v, err, shared := ct.sf.Do(ct.singleFlightKey(req), func() (interface{}, error) {
-		res, err := ct.roundTripper.RoundTrip(req)
-		if err != nil {
-			return nil, err
-		}
-		// single flight is disallowed for mutations, including file uploads
-		// hence we don't need to worry about buffering the body here
-		buf := getBuffer()
-		defer releaseBuffer(buf)
-		_, err = buf.ReadFrom(res.Body)
-		if err != nil {
-			return nil, err
-		}
-		cp := make([]byte, buf.Len())
-		copy(cp, buf.Bytes())
-		return &responseWithBody{
-			res:  res,
-			body: cp,
-		}, nil
-	})
-	if err != nil {
-		return nil, err
-	}
+	key := ct.singleFlightKey(req)
+	ct.sfMu.RLock()
+	item, shared := ct.sf[key]
+	ct.sfMu.RUnlock()
 
 	sfStats := resolve.GetSingleFlightStats(req.Context())
 	if sfStats != nil {
 		sfStats.SingleFlightUsed = true
 		sfStats.SingleFlightSharedResponse = shared
 	}
-	rwb := v.(*responseWithBody)
-	res := &http.Response{}
-	res.Status = rwb.res.Status
-	res.StatusCode = rwb.res.StatusCode
-	res.Header = rwb.res.Header.Clone()
-	res.Trailer = rwb.res.Trailer.Clone()
-	res.ContentLength = rwb.res.ContentLength
-	res.TransferEncoding = rwb.res.TransferEncoding
-	res.Close = rwb.res.Close
-	res.Uncompressed = rwb.res.Uncompressed
-	res.Request = req
+
+	if shared {
+		select {
+		case <-item.loaded:
+		case <-req.Context().Done():
+			return nil, req.Context().Err()
+		}
+
+		res := &http.Response{}
+		res.Status = item.response.Status
+		res.StatusCode = item.response.StatusCode
+		res.Header = item.response.Header.Clone()
+		res.Trailer = item.response.Trailer.Clone()
+		res.ContentLength = item.response.ContentLength
+		res.TransferEncoding = item.response.TransferEncoding
+		res.Close = item.response.Close
+		res.Uncompressed = item.response.Uncompressed
+		res.Request = req
+
+		// Restore the body
+		res.Body = io.NopCloser(bytes.NewReader(item.body))
+		return res, item.err
+	}
+
+	if sfStats != nil {
+		sfStats.SingleFlightUsed = true
+		sfStats.SingleFlightSharedResponse = false
+	}
+
+	item = &sfCacheItem{
+		loaded: make(chan struct{}),
+	}
+	ct.sfMu.Lock()
+	ct.sf[key] = item
+	ct.sfMu.Unlock()
+	defer func() {
+		close(item.loaded)
+		ct.sfMu.Lock()
+		delete(ct.sf, key)
+		ct.sfMu.Unlock()
+	}()
+
+	res, err := ct.roundTripper.RoundTrip(req)
+	if err != nil {
+		item.err = err
+		return nil, err
+	}
+
+	buf := getBuffer()
+	defer releaseBuffer(buf)
+	_, err = buf.ReadFrom(res.Body)
+	if err != nil {
+		item.err = err
+		return nil, err
+	}
+
+	item.response = res
+	item.body = make([]byte, buf.Len())
+	copy(item.body, buf.Bytes())
 
 	// Restore the body
-	res.Body = io.NopCloser(bytes.NewReader(rwb.body))
+	res.Body = io.NopCloser(bytes.NewReader(item.body))
 
 	return res, nil
 }
 
-func (ct *CustomTransport) singleFlightKey(req *http.Request) string {
+func (ct *CustomTransport) singleFlightKey(req *http.Request) uint64 {
 	keyGen := pool.Hash64.Get()
 	defer pool.Hash64.Put(keyGen)
 
@@ -283,7 +318,7 @@ func (ct *CustomTransport) singleFlightKey(req *http.Request) string {
 	}
 
 	sum := keyGen.Sum64()
-	return strconv.FormatUint(sum, 10)
+	return sum
 }
 
 type TransportFactory struct {
@@ -337,9 +372,8 @@ func (t TransportFactory) RoundTripper(enableSingleFlight bool, transport http.R
 		trace.WithPreHandler(func(r *http.Request) {
 			span := otrace.SpanFromContext(r.Context())
 			reqContext := getRequestContext(r.Context())
-			operation := reqContext.operation
 
-			commonAttributeValues := getAttributesFromOperationContext(operation)
+			var commonAttributeValues []attribute.KeyValue
 
 			subgraph := reqContext.ActiveSubgraph(r)
 			if subgraph != nil {
@@ -350,6 +384,8 @@ func (t TransportFactory) RoundTripper(enableSingleFlight bool, transport http.R
 			if attributes := baseAttributesFromContext(r.Context()); attributes != nil {
 				commonAttributeValues = append(commonAttributeValues, attributes...)
 			}
+
+			commonAttributeValues = append(commonAttributeValues, reqContext.operation.Attributes()...)
 
 			span.SetAttributes(commonAttributeValues...)